mcp subcommand cleanups
