+++ crates/chat-cli/src/cli/chat/mod.rs
@@ -4,7 +4,6 @@ mod context;
 mod conversation_state;
 mod hooks;
 mod input_source;
-pub mod mcp;
 mod message;
 mod parse;
 mod parser;
@@ -13,8 +12,8 @@ mod server_messenger;
 #[cfg(unix)]
 mod skim_integration;
 mod token_counter;
-mod tool_manager;
-mod tools;
+pub mod tool_manager;
+pub mod tools;
 pub mod util;
 
 use std::borrow::Cow;
@@ -40,11 +39,7 @@ use std::{
     fs,
 };
 
-use clap::{
-    Args,
-    Subcommand,
-    ValueEnum,
-};
+use clap::Args;
 use command::{
     Command,
     PromptsSubcommand,
@@ -351,119 +346,6 @@ impl ChatArgs {
 
         result
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Subcommand)]
-pub enum McpSubcommand {
-    /// Add or replace a configured server
-    Add(McpAdd),
-    /// Remove a server from the MCP configuration
-    #[command(alias = "rm")]
-    Remove(McpRemove),
-    /// List configured servers
-    List(McpList),
-    /// Import a server configuration from another file
-    Import(McpImport),
-    /// Get the status of a configured server
-    Status {
-        #[arg(long)]
-        name: String,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct McpAdd {
-    /// Name for the server
-    #[arg(long)]
-    pub name: String,
-    /// The command used to launch the server
-    #[arg(long)]
-    pub command: String,
-    /// Where to add the server to.
-    #[arg(long, value_enum)]
-    pub scope: Option<Scope>,
-    /// Environment variables to use when launching the server
-    #[arg(long, value_parser = parse_env_vars)]
-    pub env: Vec<HashMap<String, String>>,
-    /// Server launch timeout, in milliseconds
-    #[arg(long)]
-    pub timeout: Option<u64>,
-    /// Overwrite an existing server with the same name
-    #[arg(long, default_value_t = false)]
-    pub force: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct McpRemove {
-    #[arg(long)]
-    pub name: String,
-    #[arg(long, value_enum)]
-    pub scope: Option<Scope>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct McpList {
-    #[arg(value_enum)]
-    pub scope: Option<Scope>,
-    #[arg(long, hide = true)]
-    pub profile: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Args)]
-pub struct McpImport {
-    #[arg(long)]
-    pub file: String,
-    #[arg(value_enum)]
-    pub scope: Option<Scope>,
-    /// Overwrite an existing server with the same name
-    #[arg(long, default_value_t = false)]
-    pub force: bool,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, ValueEnum)]
-pub enum Scope {
-    Workspace,
-    Global,
-}
-
-impl std::fmt::Display for Scope {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Scope::Workspace => write!(f, "workspace"),
-            Scope::Global => write!(f, "global"),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct EnvVarParseError(String);
-
-impl std::fmt::Display for EnvVarParseError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Failed to parse environment variables: {}", self.0)
-    }
-}
-
-impl std::error::Error for EnvVarParseError {}
-
-fn parse_env_vars(arg: &str) -> Result<HashMap<String, String>, EnvVarParseError> {
-    let mut vars = HashMap::new();
-
-    for pair in arg.split(",") {
-        match pair.split_once('=') {
-            Some((key, value)) => {
-                vars.insert(key.trim().to_string(), value.trim().to_string());
-            },
-            None => {
-                return Err(EnvVarParseError(format!(
-                    "Invalid environment variable '{}'. Expected 'name=value'",
-                    pair
-                )));
-            },
-        }
-    }
-
-    Ok(vars)
 }
 
 /// Help text for the compact command
